Moved final outputprog class in a new directory
The outputprog class is now in a different directory, to be cleaner.
At this point, the outputprog class outputs a tab-separated file and
the gnuplot script for reading it, every line begins with the number of
traces and the N next columns represent the maximum value for the
Pearson coefficient of the N-th key.